requirements.txt
# Mental Health Prediction - Python Dependencies

# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Configuration
PyYAML==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.1.0
pylint==2.17.5

# Logging and Monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Optional: Advanced ML Libraries (uncomment if needed)
# xgboost==1.7.6
# lightgbm==4.0.0
# catboost==1.2

# Optional: Deep Learning (uncomment if needed)
# tensorflow==2.13.0
# torch==2.0.1

# Optional: MLOps Tools (uncomment if needed)
# mlflow==2.5.0
# dvc==3.15.0
